Refactoring affinity code
