+++ include/e2sar.hpp
@@ -2,6 +2,7 @@
 #define E2SARHPP
 
 #include "e2sarUtil.hpp"
+#include "e2sarAffinity.hpp"
 #include "e2sarCP.hpp"
 #include "e2sarDPSegmenter.hpp"
 #include "e2sarDPReassembler.hpp"
