+++ include/e2sarUtil.hpp
@@ -618,7 +618,7 @@ namespace e2sar
             {}
             // d-tor doesn't exist
             ~Optimizations() = delete;
-            
+
             static Optimizations* instance;
             static inline Optimizations* _get()
             {
@@ -627,26 +627,5 @@ namespace e2sar
                 return instance;
             }
     };
-
-    /**
-     * Set the affinity of the entire process to the cores in the vector
-     */
-    result<int> setProcessAffinity(const std::vector<int> &cores) noexcept;
-
-    /**
-     * Set calling thread affinity to specified core
-     */
-    result<int> setThreadAffinity(int core) noexcept;
-
-    /**
-     * Set calling thread affinity to exclude named cores
-     */
-    result<int> setThreadAffinityXOR(const std::vector<int> &cores) noexcept;
-
-    /**
-     * Bind process memory allocation to specified NUMA node
-     * Error out if node is invalid or NUMA not supported
-     */
-    result<int> setNUMABind(int node) noexcept;
 };
 #endif
