+++ src/meson.build
@@ -2,7 +2,7 @@ subdir('grpc')
 
 e2sar_sources = ['e2sarUtil.cpp', 'e2sarCP.cpp',
     'e2sarDPSegmenter.cpp', 'e2sarDPReassembler.cpp',
-    'e2sarNetUtil.cpp']
+    'e2sarNetUtil.cpp', 'e2sarAffinity.cpp']
 
 libe2sar_t = static_library('e2sar_t',
                         e2sar_sources,
